+++ pkg/common/constants/constants.go
@@ -23,6 +23,7 @@ const (
 	RunaiGpuLimit            = "runai-gpu-limit"
 	MpsAnnotation            = "mps"
 	StalePodgroupTimeStamp   = "kai.scheduler/stale-podgroup-timestamp"
+	LastStartTimeStamp       = "kai.scheduler/last-start-timestamp"
 
 	// Labels
 	NodePoolNameLabel        = "runai/node-pool"